User and post relations cleared

+++ controllers/feed.js
@@ -132,6 +132,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404
         throw error
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Unauthorized')
+        error.statusCode = 403
+        throw error
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl)
       }
@@ -163,13 +168,24 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404
         throw error
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Unauthorized')
+        error.statusCode = 403
+        throw error
+      }
       clearImage(post.imageUrl)
       return Post.findByIdAndRemove(postId)
     })
     .then(result => {
-      console.log(result)
+      return User.findById(req.userId)
+    })
+    .then(user => {
+      user.posts.pull(postId)
+      return user.save()
+    })
+    .then(result => {
       res.status(200).json({
-        message: 'Post deleted'
+        message: 'Post Deleted!'
       })
     })
     .catch(err => {
